clean up constants and utils
