+++ src/constants/index.js
@@ -1,0 +1,1 @@
+export const LANG_PREF_KEY = "userLanguage";
